Extracting XML reading from Model.
Also changing the definition of a Model. Now a Model is JUST one puzzles.
The puzzles management is handle by GameApplication now.
