+++ app/src/main/java/android/rushdroid/GameApplication.java
@@ -3,6 +3,9 @@ package android.rushdroid;
 import android.app.Application;
 import android.content.res.AssetManager;
 import android.rushdroid.model.Model;
+import android.rushdroid.model.ModelXML;
+import android.rushdroid.model.NotFound;
+import android.rushdroid.model.Piece;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -14,13 +17,23 @@ import java.util.List;
  */
 public class GameApplication extends Application {
   private Model game;
+  private List<Model> games;
+
+  public void selectPuzzle(int i) throws NotFound {
+    try {
+      game = games.get(i);
+    } catch (IndexOutOfBoundsException e) {
+      throw new NotFound("Invalid id greater than puzzles number.");
+    }
+  }
+
+  public int size() { return games.size(); }
 
   @Override
   public void onCreate() {
     AssetManager am = this.getApplicationContext().getAssets();
     try {
-      InputStream in = am.open("rushpuzzles.xml");
-      this.game = new Model(in);
+      this.games = new ModelXML().read(am.open("rushpuzzles.xml"));
     } catch (IOException e) {
       e.printStackTrace();
     }
